todas las demas

+++ src/main/java/com/example/parcial/demo/repositories/PracticaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Date;
 import java.util.List;
 
-public interface PracticaRepository extends JpaRepository<Practica, Long> {
+public interface PracticaRepository extends JpaRepository<Practica, Integer> {
     List<Practica> findByDocente_Identificacion(String identificacion);
     List<Practica> findByFechaSalidaBetween(Date inicio, Date fin);
 }
